ci: add --allow-env for deno test
./test/frontmatter_test.ts (uncaught error)
769
error: (in promise) NotCapable: Requires env access to "__MINIMATCH_TESTING_PLATFORM__", run again with the --allow-env flag
770
    at Object.getEnv [as get] (ext:deno_os/30_os.js:124:10)
771
    at denoEnvGet (ext:deno_node/_process/process.ts:27:21)
772
    at Object.get (ext:deno_node/_process/process.ts:44:22)
773
    at file:///home/runner/.cache/deno/npm/registry.npmjs.org/minimatch/10.0.1/dist/esm/index.js:69:21
774
This error was not caught from a test and caused the test runner to fail on the referenced module.
775
It most likely originated from a dangling promise, event/timeout handler or top-level code.